refactor(basecheckbox): improve typing on BaseCheckbox, Checkbox and Radio components
