+++ packages/components/forms/src/base-checkbox/index.ts
@@ -1,2 +1,2 @@
 export { BaseCheckbox } from './BaseCheckbox';
-export type { BaseCheckboxProps } from './types';
+export type { BaseCheckboxProps } from './BaseCheckbox';
